Crate ckb_types

The Core Types Library

This Library provides the essential types for building ckb.

Re-exports

pub use bytes;

Modules

constants
All Constants.

core
Rust types.

error

packed

prelude

utilities

Macros

h160

h256

u256
A macro used to construct a fixed uint in compile time.

Structs

H160

H256

U128
Fixed non-negative integer type.

U256
Fixed non-negative integer type.