Module ckb_types::core

Rust types.

Packed bytes wrappers are not enough for all usage scenarios.

Modules

cell

error

service

Macros

capacity_bytes

Structs

BlockBuilder

BlockEconomicState

BlockExt

BlockIssuance

BlockReward

BlockView

Capacity

EpochExt

EpochNumberWithFraction
Represents an epoch number with a fraction unit, it can be used to accurately represent the position for a block within an epoch.

HeaderBuilder

HeaderView

MinerReward

Ratio

TransactionBuilder

TransactionInfo

TransactionMeta

TransactionMetaBuilder

TransactionView

UncleBlockVecView

UncleBlockView

Enums

DepType

ScriptHashType

Type Definitions

BlockNumber

CapacityResult

Cycle

EpochNumber

PublicKey

Version